Fading the "Scroll to Top" button in and out
Previously the "scroll to Top"-button was always shown.
In my opinion this makes little sense when you are already at the top
and may even lead to confusion (nothing happens when you press the button,
when at top).
So this change will fade the button in and out depending on where you are.
It is leveraging the already implemented scrollPage function.

Signed-off-by: Simon Egli <deadolus@gmail.com>