add script section in package.json file
Add a script section in package.json file so that we can use the node script feature. This allow to use the embedded version of gulp natively.

gulp is installed locally due to its mention in package.json file, it should be easy to use gulp without to type the whole path. Using the node script feature, it is now possible to use the following commands :

```bash
npm run build // to run the *build* task of gulp
npm run dev // to run the *dev* task of gulp
```